Extension variable for docker remote api version.
Provide a way to set the docker remote api version on the underlying docker-java library.
The default version seems to be UNKNOWN_VERSION, which usually is ok, but if you are
talking to a v2 registry, the api version is used as a proxy to determine how to send
authentication headers.  If unset, v2 registry authentication will fail.  See also
https://github.com/docker-java/docker-java/issues/506