Introduce explicit declaration that task supports onNext callback
Before it was unclear if the logic in onNext is actually evaluated by a task without having a look at the code. Task now express this automatically if they implement the trait ResultCallback.

Furthermore, simplifies the inheritance hierarchy by removing the unnecessary parent class. All tasks of this plugin had the same handling anyway.