Add property buildLabels to add build info to image
Adding labels into this property will not invalidate the cache key.
It might be useful to add manifest information without trigerring
a new build every time.
But whenb a build is triggered, the information goes in.

Add a test validating that even when buildLabels change, it does
not trigger a rebuild of the image. When a rebuild is triggered,
new buildLabels will be included.