Fix DockerBuildImageFunctionalTest failing when re-executing them
When re-executing the tests for the seconds time, the test "task not
up-to-date when image not in registry" was failing with the following
error:

Task :removeImage FAILED
Removing image with ID 'd90aa9bfe0e6'.

FAILURE: Build failed with an exception.

* What went wrong:
Execution failed for task ':removeImage'.
> {"message":"conflict: unable to delete d90aa9bfe0e6 (cannot be forced) - image has dependent child images"}

The problem was that the tests are interfering with each other because
the images created were not unique. Make sure the image definition
created by imageCreation() method will be unique by adding a layer which
echo a UUID.