ISSUE-834: Add multiple tags support application plugins
Only being able to specify one tag for these projects when you can
specify multiple in the DockerBuildImage tasks is a real drag. It means
you would have to add a custom DockerBuildImage or DockerTagImage task
when the whole point of the plugin is to abstract all that out.

Tested by:
    `./gradlew clean build`