Upgrade docker-java to fix symlink issue
When the DockerBuildImage input dir contained symlinks, the symlinks
were not sent to the docker daemon. This caused docker build to fail or
succeed but with missing symlinks depending if the DockerFile uses the
expected symlinks directly of if it uses a folder containing the symlinks.

The docker-java version 3.1.4 fixes this issue and now the symbolic
links are sent as is to the docker daemon, same as what native docker
client does.

Also adjust one test to the new behavior in 3.1.4. The error message
when a container does not exist used to be:
{"message":"No such container: idonotexist"}
and now is :
No such container: idonotexist

Resolves #837