Make it invalid to confirm previously nullified properties
In previous versions, the potentially configured values were nullified before executing the task.
Since changing a `Property` once the task has started is illegal in Gradle 6.0, we now make the
properties final upon task creation.
This will clearly indicate to users that they cannot be configured on this task.

Issue #868