Allow configuration of `DockerExistingImage#targetImageId` with a build task.
Based on discussion #977.

This is like a convenience functionality where the image ID of
`DockerBuildImage` task will be used if the image was built in the same Gradle
build, otherwise the first image name/tag will be used from
`DockerBuildImage#images` (assuming it exists on the local filesystem).

I believe that this kind of auto-configuration should fit well for the majority
of the use cases.